Tracking/SkyLines/Client: support negative GeoPoint angles
The FromBE32() call dropped the "signed", and this commit re-adds an
explicit cast to signed integer to restore the sign.  Fixes TRAC
#3601.

+++ src/Tracking/SkyLines/Client.cpp
@@ -135,7 +135,7 @@ SkyLinesTracking::Client::SendUserNameRequest(uint32_t user_id)
 static constexpr Angle
 ImportAngle(int32_t src)
 {
-  return Angle::Degrees(fixed(FromBE32(src)) / 1000000);
+  return Angle::Degrees(fixed(int32_t(FromBE32(src))) / 1000000);
 }
 
 /**
